refactor: move Locale to shared, type preferences locale, remove gateway
- Move `Locale` enum from `kartoteka-i18n` to `kartoteka-shared` so domain and shared DTOs can use it without circular deps
- `domain::preferences::Preferences.locale` and `UpdatePreferencesRequest.locale` are now typed `Locale` instead of `String` — invalid locales (e.g. "ja") are rejected at deserialization (422)
- `shared::dto::responses` updated to use `Locale` for `PreferencesResponse` and `UpdatePreferencesBody`
- `kartoteka-i18n` reduced to a re-export shim; `kartoteka-frontend` no longer depends on it
- Remove `gateway/` Cloudflare Worker (deprecated, replaced by Rust server MCP): justfile recipes, dependabot npm entry, CodeQL JS/TS job
