+++ .github/dependabot.yml
@@ -27,7 +27,7 @@ updates:
         update-types: ["version-update:semver-major"]
 
   - package-ecosystem: "npm"
-    directory: "/gateway"
+    directory: "/crates/frontend"
     schedule:
       interval: "weekly"
       day: "monday"
