+++ .github/workflows/codeql.yml
@@ -27,9 +27,6 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - language: javascript-typescript
-            build-mode: none
-            paths: gateway
           - language: actions
             build-mode: none
             paths: .github
