+++ crates/frontend/Cargo.toml
@@ -12,7 +12,6 @@ workspace = true
 
 [dependencies]
 kartoteka-shared = { path = "../shared" }
-kartoteka-i18n = { path = "../i18n" }
 leptos = { workspace = true }
 leptos_router = { workspace = true }
 leptos_meta = { workspace = true, optional = true }
@@ -47,7 +46,7 @@ chrono-tz = { workspace = true, optional = true }
 tokio = { workspace = true, optional = true }
 
 [package.metadata.cargo-machete]
-ignored = ["axum", "js-sys", "kartoteka-i18n", "leptos_meta"]
+ignored = ["axum", "js-sys", "leptos_meta"]
 
 [features]
 default = []
