+++ crates/shared/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod auth_ctx;
 pub mod constants;
 pub mod date_utils;
+pub mod locale;
+pub use locale::Locale;
 #[allow(dead_code)]
 pub(crate) mod deserializers;
 pub mod dto;
